AI Astrology/requirements.txt
flask>=3.0
python-dateutil>=2.9
gunicorn>=21.2
